# Clientory App/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{id}/prompt-sets/generate:
    post:
      operationId: generatePromptSet
      tags: [businesses]
      summary: Generate a locked prompt set for a firm
      description: Generates 10 prompts across 5 categories. Idempotent — returns existing set if already generated.
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: New prompt set generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptSet"
        "200":
          description: Existing prompt set returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptSet"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/prompt-sets/current:
    get:
      operationId: getPromptSet
      tags: [businesses]
      summary: Get the current locked prompt set for a firm
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current prompt set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptSet"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No prompt set found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/prompt-sets/regenerate:
    post:
      operationId: regeneratePromptSet
      tags: [businesses]
      summary: Archive current prompt set and generate a new one
      description: Only allowed if the firm profile has been updated since the last generation.
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: New prompt set generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptSet"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Profile not updated since last generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PromptItem:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        text:
          type: string
        sourceVariables:
          type: object
          additionalProperties: true
        audience:
          type: string
          enum: [individual, business]
          nullable: true
      required:
        - id
        - category
        - text
        - sourceVariables
    PromptSet:
      type: object
      properties:
        id:
          type: string
        businessId:
          type: string
        generatedAt:
          type: string
          format: date-time
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/PromptItem"
        isLocked:
          type: boolean
        substitutionNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - generatedAt
        - prompts
        - isLocked
        - createdAt
...